Refactor to resolve bug allowing changing isheader and expiration after file uploaded
