+++ src/pages/UploadPage/UploadPage.js
@@ -5,6 +5,7 @@ import Filedrop from '../../components/Filedrop/Filedrop';
 import gql from 'graphql-tag';
 import { Mutation } from 'react-apollo';
 import { Redirect } from 'react-router-dom';
+import Papa from 'papaparse';
 
 const UPLOAD_SHEET = gql`
 	mutation UPLOAD_SHEET($sheetData: JSON!, $expireIn: Int!) {
@@ -23,41 +24,11 @@ const UploadPage = () => {
 	const [expireIn, setExpireIn] = useState(72);
 	const [header, setHeader] = useState(true);
 	const [disableSubmit, setDisableSubmit] = useState(true);
-	const [sheetData, setSheetData] = useState();
 	const [redirect, setRedirect] = useState(false);
-
-	/**
-	 * @function
-	 * Set expireIn state, in Hrs
-	 * @param {object} e - dom event
-	 */
-	const _handleChange = e => {
-		e.preventDefault();
-		setExpireIn(e.target.value);
-	};
-
-	/**
-	 * @function
-	 * Toggle first row of CSV header boolean
-	 * @param {object} e - dom event
-	 */
-	const _toggleHeader = e => {
-		e.preventDefault();
-		setHeader(!header);
-	};
-
-	/**
-	 * @function
-	 * Handle response from <FileDrop />
-	 * @param {object} response - converted data from FileDrop
-	 */
-	const handleJSONData = response => {
-		setSheetData(response.data);
-		setDisableSubmit(false);
-	};
+	const [file, setFile] = useState();
 
 	const onCompleted = data => {
-		const id = data.createSheet._id;
+		// const id = data.createSheet._id;
 		setRedirect(true);
 	};
 
@@ -68,26 +39,33 @@ const UploadPage = () => {
 		}
 	};
 
+	const onDrop = (file, rejectedFile) => {
+		if (
+			file[0] &&
+			file[0].name &&
+			file[0].name.indexOf('.csv') + 4 === file[0].name.length
+		) {
+			setFile(file);
+
+			setDisableSubmit(false);
+		} else {
+			// TODO handle Error on UI
+			console.error(
+				`file rejected!: \n ${rejectedFile[0].name}\n ${rejectedFile[0].type}`
+			);
+		}
+	};
+
 	return (
 		<StyledDiv>
-			<Filedrop firstRowHeader={header} handleJSONData={handleJSONData} />
+			<Filedrop
+				firstRowHeader={header}
+				expireIn={expireIn}
+				setHeader={setHeader}
+				setExpireIn={setExpireIn}
+				onDrop={onDrop}
+			/>
 			<StyledForm disableSubmit={disableSubmit}>
-				<Options>
-					<label>
-						Expires In:
-						<select value={expireIn} onChange={_handleChange}>
-							<option value={4}>4 Hours</option>
-							<option value={8}>8 Hours</option>
-							<option value={24}>1 Day</option>
-							<option value={72}>3 Days</option>
-							<option value={120}>5 Days</option>
-						</select>
-					</label>
-					<HeaderToggle active={header}>
-						<span>First Row Header:</span>
-						<button onClick={_toggleHeader}>{header ? 'Yes' : 'No'}</button>
-					</HeaderToggle>
-				</Options>
 				<Mutation mutation={UPLOAD_SHEET} onCompleted={onCompleted}>
 					{(uploadSheet, { loading, error, data }) => (
 						<div>
@@ -96,8 +74,17 @@ const UploadPage = () => {
 								type="submit"
 								onClick={e => {
 									e.preventDefault();
-									uploadSheet({
-										variables: { sheetData, expireIn: parseInt(expireIn) },
+									Papa.parse(file[0], {
+										header: header,
+										download: true,
+										skipEmptyLines: false,
+										complete: ({ data }) =>
+											uploadSheet({
+												variables: {
+													sheetData: data,
+													expireIn: parseInt(expireIn),
+												},
+											}),
 									});
 								}}
 								value="Upload File"
@@ -163,36 +150,6 @@ const StyledForm = styled.form`
 				props.disableSubmit ? '' : props.theme.boxShadowDark};
 		}
 	}
-`;
-
-const HeaderToggle = styled.label`
-	margin-left: 2rem;
-
-	button {
-		color: white;
-		cursor: pointer;
-		outline: none;
-		background: ${props =>
-			props.active
-				? props.theme.gradient.greenBlue
-				: props.theme.color.backgroundDarkest};
-		box-shadow: ${props => (props.active ? props.theme.boxShadow : '')};
-		border: none;
-		font-family: ${props => props.theme.font.main};
-		font-size: 1rem;
-		border-radius: 4px;
-		margin-left: 1rem;
-		padding-top: 5px;
-		padding-bottom: 5px;
-		width: 70px;
-	}
-`;
-
-const Options = styled.div`
-	background-color: ${props => props.theme.color.backgroundDark};
-	border-radius: 3px;
-	padding-top: 1rem;
-	padding-bottom: 1rem;
 `;
 
 export default UploadPage;
